bench/requirements.txt
locust>=2.29,<3
locust-plugins>=4.5,<5
toxiproxy-python>=0.1,<1
websocket-client>=1.7,<2
requests>=2.31,<3
prometheus-api-client>=0.5,<1
